BACKOFFICE navigation and quick fixes
