--- src/Pyz/Zed/Faq/Communication/Controller/ListController.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace Pyz\Zed\Faq\Communication\Controller;
-
-use Pyz\Zed\Faq\Communication\FaqCommunicationFactory;
-use Spryker\Zed\Kernel\Communication\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
-
-/**
- * @method FaqCommunicationFactory getFactory()
- */
-class ListController extends AbstractController
-{
-    public function indexAction(): array
-    {
-        $table = $this->getFactory()->createFaqTable();
-        return $this->viewResponse([
-            'questionTable' => $table->render()
-        ]);
-    }
-
-    public function tableAction(): JsonResponse {
-        $table = $this->getFactory()->createFaqTable();
-
-        return $this->jsonResponse($table->fetchData());
-    }
-}
