+++ src/Pyz/Zed/Faq/Presentation/Edit/index.twig
@@ -1,6 +1,6 @@
 {% extends '@Gui/Layout/layout.twig' %}
 
-{% set widget_title = 'Create a new planet' %}
+{% set widget_title = 'Edit question' %}
 
 {% block head_title widget_title %}
 {% block section_title widget_title %}
