+++ src/Pyz/Zed/Faq/Presentation/Index/index.twig
@@ -1,20 +1,15 @@
 {% extends '@Gui/Layout/layout.twig' %}
-
-{% set widget_title = 'Create a new planet' %}
-
+{% set widget_title = 'FAQ' %}
 {% block head_title widget_title %}
 {% block section_title widget_title %}
-
+{% block action %}
+{#    {{ createActionButton('/planet-star/list', 'See Star' | trans) }}#}
+{#    {{ createActionButton('/planet/create', 'Create Planet' | trans) }}#}
+{% endblock %}
 {% block content %}
-
-    {% embed '@Gui/Partials/widget.twig' %}
-
+    {% embed '@Gui/Partials/widget.twig' with { widget_title: 'List of question' } %}
         {% block widget_content %}
-
-            {{ form(planetForm) }}
-
+            {{ questionTable | raw }}
         {% endblock %}
-
     {% endembed %}
-
 {% endblock %}
