--- src/Pyz/Zed/Faq/Presentation/List/index.twig
@@ -1,15 +1,0 @@
-{% extends '@Gui/Layout/layout.twig' %}
-{% set widget_title = 'FAQ' %}
-{% block head_title widget_title %}
-{% block section_title widget_title %}
-{% block action %}
-{#    {{ createActionButton('/planet-star/list', 'See Star' | trans) }}#}
-{#    {{ createActionButton('/planet/create', 'Create Planet' | trans) }}#}
-{% endblock %}
-{% block content %}
-    {% embed '@Gui/Partials/widget.twig' with { widget_title: 'List of question' } %}
-        {% block widget_content %}
-            {{ questionTable | raw }}
-        {% endblock %}
-    {% endembed %}
-{% endblock %}
